# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Absher Travel & Tourism API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /app-images:
    get:
      operationId: listAppImages
      tags: [app-images]
      summary: List active app images within their schedule window, ordered by sortOrder
      parameters:
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of active app images
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AppImage"
    post:
      operationId: createAppImage
      tags: [app-images]
      summary: Create an app image (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppImageInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppImage"
  /app-images/admin/list:
    get:
      operationId: listAppImagesAdmin
      tags: [app-images]
      summary: List all app images (admin, includes inactive/out-of-window)
      parameters:
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of all app images
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AppImage"
  /app-images/{id}:
    patch:
      operationId: updateAppImage
      tags: [app-images]
      summary: Update an app image (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppImageUpdate"
      responses:
        "200":
          description: Updated app image
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppImage"
    delete:
      operationId: deleteAppImage
      tags: [app-images]
      summary: Delete an app image (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List the authenticated user's notifications
      parameters:
        - name: unreadOnly
          in: query
          schema: { type: boolean }
      responses:
        "200":
          description: Notification list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
...
  /notifications/{id}:
    delete:
      operationId: deleteNotification
      tags: [notifications]
      summary: Delete one of the authenticated user's notifications
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                type: object
                required: [deleted]
                properties:
                  deleted: { type: integer }
  /notifications/send:
    post:
      operationId: sendNotification
      tags: [notifications]
      security:
        - bearerAuth: []
      summary: Admin — create and push a notification to users (requires 'notifications' permission)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NotificationSendRequest"
      responses:
        "200":
          description: Sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationSendResponse"
  /notifications/admin/history:
    get:
      operationId: listAdminNotificationHistory
      tags: [notifications]
      security:
        - bearerAuth: []
      summary: Admin — recent admin-sent notifications (requires 'notifications' permission)
      responses:
        "200":
          description: Admin-sent notification history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"

  # ── PUSH TOKENS ──────────────────────────────────────────────────────────
...
components:
  schemas:
...
    AppImage:
      type: object
      required: [id, category, imageUrl, sortOrder, isActive, createdAt, updatedAt]
      properties:
        id: { type: integer }
        category: { type: string }
        titleAr: { type: ["string", "null"] }
        titleEn: { type: ["string", "null"] }
        imageUrl: { type: string }
        linkUrl: { type: ["string", "null"] }
        relatedEntityType: { type: ["string", "null"] }
        relatedEntityId: { type: ["string", "null"] }
        sortOrder: { type: integer }
        startDate: { type: ["string", "null"] }
        endDate: { type: ["string", "null"] }
        isActive: { type: boolean }
        createdAt: { type: string }
        updatedAt: { type: string }
    AppImageInput:
      type: object
      required: [category, imageUrl]
      properties:
        category: { type: string, minLength: 1 }
        titleAr: { type: ["string", "null"] }
        titleEn: { type: ["string", "null"] }
        imageUrl: { type: string, minLength: 1 }
        linkUrl: { type: ["string", "null"] }
        relatedEntityType: { type: ["string", "null"] }
        relatedEntityId: { type: ["string", "null"] }
        sortOrder: { type: integer }
        startDate: { type: ["string", "null"] }
        endDate: { type: ["string", "null"] }
        isActive: { type: boolean }
    AppImageUpdate:
      type: object
      properties:
        category: { type: string, minLength: 1 }
        titleAr: { type: ["string", "null"] }
        titleEn: { type: ["string", "null"] }
        imageUrl: { type: string, minLength: 1 }
        linkUrl: { type: ["string", "null"] }
        relatedEntityType: { type: ["string", "null"] }
        relatedEntityId: { type: ["string", "null"] }
        sortOrder: { type: integer }
        startDate: { type: ["string", "null"] }
        endDate: { type: ["string", "null"] }
        isActive: { type: boolean }
...
    Notification:
      type: object
      required: [id, userId, titleAr, titleEn, messageAr, messageEn, channel, isRead, createdAt]
      properties:
        id: { type: string }
        userId: { type: string }
        titleAr: { type: string }
        titleEn: { type: string }
        messageAr: { type: string }
        messageEn: { type: string }
        channel:
          type: string
          enum: [in_app, email, whatsapp, sms]
        relatedEntityType: { type: ["string", "null"] }
        relatedEntityId: { type: ["string", "null"] }
        url: { type: ["string", "null"] }
        imageUrl: { type: ["string", "null"] }
        isRead: { type: boolean }
        sentBy: { type: ["string", "null"] }
        createdAt: { type: string }
...
    NotificationSendRequest:
      type: object
      required: [titleAr, titleEn, messageAr, messageEn, audience]
      properties:
        titleAr: { type: string }
        titleEn: { type: string }
        messageAr: { type: string }
        messageEn: { type: string }
        audience: { type: string, enum: [all, users, group] }
        userIds:
          type: array
          items: { type: string }
        roles:
          type: array
          items: { type: string, enum: [customer, agent, admin, super_admin] }
        url: { type: string }
        imageUrl: { type: string }
    NotificationSendResponse:
      type: object
      required: [sentCount]
      properties:
        sentCount: { type: integer }
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT